+++ embedded/common/cross_door.rs
@@ -7,7 +7,6 @@
 use alloc::sync::Arc;
 use ariel_os::time::{Duration, Instant};
 use behaviortree::prelude::*;
-use spin::Mutex;
 
 /// Returns `true` if sleeping, `false` after sleep time ended.
 #[must_use]